Updated Nav, Signup and Login
